+++ languagetool-language-modules/ar/src/main/java/org/languagetool/language/Arabic.java
@@ -41,7 +41,21 @@ import java.util.*;
  */
 public class Arabic extends Language implements AutoCloseable {
 
-  private static final Language DEFAULT_ARABIC = new AlgerianArabic();
+
+  public static final String TASHKEEL_CHARS =
+    "\u064B"    // Fathatan
+    + "\u064C"  // Dammatan
+    + "\u064D"  // Kasratan
+    + "\u064E"  // Fatha
+    + "\u064F"  // Damma
+    + "\u0650"  // Kasra
+    + "\u0651"  // Shadda
+    + "\u0652"  // Sukun
+    + "\u0653"  // Maddah Above
+    + "\u0654"  // Hamza Above
+    + "\u0655"  // Hamza Below
+    + "\u0656"  // Subscript Alef
+    + "\u0640"; // Tatweel
 
   private LanguageModel languageModel;
 
@@ -58,11 +72,6 @@ public class Arabic extends Language implements AutoCloseable {
   @Override
   public String[] getCountries() {
     return new String[]{"", "SA", "DZ", "BH", "EG", "IQ", "JO", "KW", "LB", "LY", "MA", "OM", "QA", "SD", "SY", "TN", "AE", "YE"};
-  }
-
-  @Override
-  public Language getDefaultLanguageVariant() {
-    return DEFAULT_ARABIC;
   }
 
   @Override
@@ -95,9 +104,7 @@ public class Arabic extends Language implements AutoCloseable {
   public Contributor[] getMaintainers() {
     return new Contributor[]{
       new Contributor("Taha Zerrouki"),
-      new Contributor("Sohaib Afifi"),
-      new Contributor("Imen Kali"),
-      new Contributor("Karima Tchoketch"),
+      new Contributor("Sohaib Afifi")
     };
   }
 
@@ -111,18 +118,23 @@ public class Arabic extends Language implements AutoCloseable {
         Arrays.asList("]", ")", "}", "»", "﴿", "\"", "'")),
 
       // specific to Arabic :
-      new ArabicHunspellSpellerRule(messages, this, userConfig, altLanguages),
-      //new MorfologikArabicSpellerRule(messages, this),
+      new ArabicHunspellSpellerRule(messages, userConfig),
       new ArabicCommaWhitespaceRule(messages),
       new ArabicDoublePunctuationRule(messages),
       new LongSentenceRule(messages, userConfig, -1, false),
-      new ArabicWordRepeatRule(messages, this),
-      new ArabicSimpleReplaceRule(messages, this)
+      new ArabicWordRepeatRule(messages),
+      new ArabicSimpleReplaceRule(messages),
+      new ArabicDiacriticsRule(messages),
+      new ArabicHomophonesRule(messages),
+      new ArabicRedundancyRule(messages),
+      new ArabicWordCoherencyRule(messages),
+      new ArabicWordinessRule(messages),
+      new ArabicWrongWordInContextRule(messages)
     );
   }
 
   @Override
-  public List<Rule> getRelevantLanguageModelRules(ResourceBundle messages, LanguageModel languageModel, UserConfig userConfig) throws IOException {
+  public List<Rule> getRelevantLanguageModelRules(ResourceBundle messages, LanguageModel languageModel, UserConfig userConfig) {
     return Arrays.asList(
       new ArabicConfusionProbabilityRule(messages, languageModel, this)
     );
@@ -134,13 +146,13 @@ public class Arabic extends Language implements AutoCloseable {
   }
 
   @Override
-  public synchronized LanguageModel getLanguageModel(File indexDir) throws IOException {
+  public synchronized LanguageModel getLanguageModel(File indexDir) {
     languageModel = initLanguageModel(indexDir, languageModel);
     return languageModel;
   }
 
   @Override
-  public void close() throws Exception {
+  public void close() {
     if (languageModel != null) {
       languageModel.close();
     }